fix DistanceRewardCalculator and make distance value into float type

+++ mazerunner/utils.py
@@ -43,7 +43,7 @@ def fix_at_file(cmd, testcase):
 
 def get_distance_from_fn(filename):
     match = re.search(r'dis:(\d+)', filename)
-    return None if not match else int(match.group(1))
+    return None if not match else float(match.group(1))
 
 def run_command(cmd, testcase):
     cmd, stdin = fix_at_file(cmd, testcase)
